events: Reorganize tests
Move tests to the proper files, streamline test names and remove
duplicates.
