+++ crates/ruma-common/tests/events/file.rs
@@ -9,10 +9,7 @@ use ruma_common::{
         file::{EncryptedContentInit, FileContentInfo, FileEventContent},
         message::MessageContent,
         relation::InReplyTo,
-        room::{
-            message::{FileMessageEventContent, MessageType, Relation, RoomMessageEventContent},
-            EncryptedFileInit, JsonWebKeyInit, MediaSource,
-        },
+        room::{message::Relation, JsonWebKeyInit},
         AnyMessageLikeEvent, MessageLikeEvent,
     },
     mxc_uri,
@@ -216,117 +213,4 @@ fn message_event_deserialization() {
     assert_eq!(message_event.room_id, "!roomid:notareal.hs");
     assert_eq!(message_event.sender, "@user:notareal.hs");
     assert!(message_event.unsigned.is_empty());
-}
-
-#[test]
-fn room_message_plain_content_serialization() {
-    let message_event_content =
-        RoomMessageEventContent::new(MessageType::File(FileMessageEventContent::plain(
-            "Upload: my_file.txt".to_owned(),
-            mxc_uri!("mxc://notareal.hs/file").to_owned(),
-            None,
-        )));
-
-    assert_eq!(
-        to_json_value(&message_event_content).unwrap(),
-        json!({
-            "body": "Upload: my_file.txt",
-            "url": "mxc://notareal.hs/file",
-            "msgtype": "m.file",
-        })
-    );
-}
-
-#[test]
-fn room_message_encrypted_content_serialization() {
-    let message_event_content =
-        RoomMessageEventContent::new(MessageType::File(FileMessageEventContent::encrypted(
-            "Upload: my_file.txt".to_owned(),
-            EncryptedFileInit {
-                url: mxc_uri!("mxc://notareal.hs/file").to_owned(),
-                key: JsonWebKeyInit {
-                    kty: "oct".to_owned(),
-                    key_ops: vec!["encrypt".to_owned(), "decrypt".to_owned()],
-                    alg: "A256CTR".to_owned(),
-                    k: Base64::parse("TLlG_OpX807zzQuuwv4QZGJ21_u7weemFGYJFszMn9A").unwrap(),
-                    ext: true,
-                }
-                .into(),
-                iv: Base64::parse("S22dq3NAX8wAAAAAAAAAAA").unwrap(),
-                hashes: [(
-                    "sha256".to_owned(),
-                    Base64::parse("aWOHudBnDkJ9IwaR1Nd8XKoI7DOrqDTwt6xDPfVGN6Q").unwrap(),
-                )]
-                .into(),
-                v: "v2".to_owned(),
-            }
-            .into(),
-        )));
-
-    assert_eq!(
-        to_json_value(&message_event_content).unwrap(),
-        json!({
-            "body": "Upload: my_file.txt",
-            "file": {
-                "url": "mxc://notareal.hs/file",
-                "key": {
-                    "kty": "oct",
-                    "key_ops": ["encrypt", "decrypt"],
-                    "alg": "A256CTR",
-                    "k": "TLlG_OpX807zzQuuwv4QZGJ21_u7weemFGYJFszMn9A",
-                    "ext": true
-                },
-                "iv": "S22dq3NAX8wAAAAAAAAAAA",
-                "hashes": {
-                    "sha256": "aWOHudBnDkJ9IwaR1Nd8XKoI7DOrqDTwt6xDPfVGN6Q"
-                },
-                "v": "v2",
-            },
-            "msgtype": "m.file",
-        })
-    );
-}
-
-#[test]
-fn room_message_plain_content_deserialization() {
-    let json_data = json!({
-        "body": "Upload: my_file.txt",
-        "url": "mxc://notareal.hs/file",
-        "msgtype": "m.file",
-    });
-
-    let event_content = from_json_value::<RoomMessageEventContent>(json_data).unwrap();
-    let content = assert_matches!(event_content.msgtype, MessageType::File(content) => content);
-    assert_eq!(content.body, "Upload: my_file.txt");
-    let url = assert_matches!(content.source, MediaSource::Plain(url) => url);
-    assert_eq!(url, "mxc://notareal.hs/file");
-}
-
-#[test]
-fn room_message_encrypted_content_deserialization() {
-    let json_data = json!({
-        "body": "Upload: my_file.txt",
-        "file": {
-            "url": "mxc://notareal.hs/file",
-            "key": {
-                "kty": "oct",
-                "key_ops": ["encrypt", "decrypt"],
-                "alg": "A256CTR",
-                "k": "TLlG_OpX807zzQuuwv4QZGJ21_u7weemFGYJFszMn9A",
-                "ext": true
-            },
-            "iv": "S22dq3NAX8wAAAAAAAAAAA",
-            "hashes": {
-                "sha256": "aWOHudBnDkJ9IwaR1Nd8XKoI7DOrqDTwt6xDPfVGN6Q"
-            },
-            "v": "v2",
-        },
-        "msgtype": "m.file",
-    });
-
-    let event_content = from_json_value::<RoomMessageEventContent>(json_data).unwrap();
-    let content = assert_matches!(event_content.msgtype, MessageType::File(content) => content);
-    assert_eq!(content.body, "Upload: my_file.txt");
-    let encrypted_file = assert_matches!(content.source, MediaSource::Encrypted(f) => f);
-    assert_eq!(encrypted_file.url, "mxc://notareal.hs/file");
 }
